requirements.txt
numpy
pandas
Pillow
scikit-learn
torch
torchvision
tqdm
transformers
